issue #71: create a terse presenter
The ColorConsolePresenter always prints an empty line before a
scenario. That behaviour was removed here, as we don't want a newline
in every case.
I believe that we have to add a method for after a scenario, like we do
with a step result. Then we could simplify the TerseRealTimePresenter
as well.